EUDEMO first wall panelling (fusion-power-plant-framework/bluemira#2060)
* Add WrappedString geometry parameterisation

This shape is for the panels within the eudemo IVC

* Replace 'WrappedString' geometry with function

* Add replication of BLUEPRINT's Paneller

* Fix circular import & add eudemo panelling test

* Add some (failing) tests for PanellingDesigner

It was discussed offline that this way of defining the optimisation
problem was not the best. In reality, the maximum length of a panel does
not need to be constrained. It was discussed that we will try
constraining the minimum length and the angle between the panels.

The existing optimisation problem was quite picky, and it was difficult
to find a solution that did not violate constraints for a given min/max
panel length.

* --wip-- [skip ci]

* Add a couple more panelling designer tests

* Clean up file organisation in eudemo.ivc.panelling

* Fix flake8 issues in eudemo.ivc.panelling

* Move optimisation functions out of Paneller

These belong in the PanellingOptProblem class.

* Extract boundary interpolation from Paneller

* Fix incorrect panel length calculation

* Simplify calculation in Paneller

* Remove deprecated panelling implementation

We switched to constraining min-length and max-angle, from min-length
and max-length.

* Remove empty file

* Try increasing no. of panels if optimisation fails

* Fix running run/mock mode in test

* Raise error if no feasible solution to panelling

* Fix flake8 warnings

* Resolve TODOs with comments

* Remove BLUEPRINT comment that makes no sense

Out of the context of BLUEPRINT, this comment is more confusing than
helpful.

* Remove other TODOs/WONTDOs

* Improve docstring

* Fix long line flake8 warning

* Return initial guess in paneller if design fails

Rather than throw an error, return the initial guess (which is often a
decent attempt), and print a warning.

* Avoid division by zero in Paneller

* Catch ExternalOpt error in panelling optimisation

* Remove unused variable in 'make_pivoted_string'

* Attempt to fix UnboundLocalError

The test is passing on my machine, so not sure if this will work or not.

* Attempt to fix panelling test

The test was passing on my machine, but failing on the CI with an NLOpt
error. The solution is to catch those NLOpt errors, and retry the
optimisation or return a fall back.

* Make sure plots are close in panelling tests

* Default warn to true when opt checking constraints

* Tidy array formatting in paneller test

* Use `.xz` not indexing for coordinate retrieval

* Remove use of 'out' argument in numpy calculations

No need for the performance optimisation here in the end. It's not used
in an optimisation loop.

* Remove unused array in make_pivoted_string

Also use `np.deg2rad` not a constant, for angular unit conversion

* Remove duplicated length norm function

* Remove duplication of geometry helpers

* Remove use of 'out' in numpy operation

* Add missing return type annotation

* Add tests for panelling in/outboard separately

This is how the panelling designer will be used for EUDEMO.

* Rename shadowed builtin variable name

* Refine boundary discretisation when panelling

The test `test_panels_fully_enclose_wall_boundary` was failing on some
machines, due to a cross over between the panelling boundary and the
generated panels. This cross-over was down to a poor interpolation of
the boundary around a sharp (ish) corner.

We can fix this test, and avoid similar issues appearing in production,
by increasing the discretisation of the boundary before we interpolate.
This should not effect performance much, as we're not doing massively
expensive operations with the boundary (e.g., keep-out-zone
constraints).

We may question why we're discretising and then interpolating a wire,
but we don't currently have a `BluemiraWire.tangent_at` method. We may
consider implementing this in the future, at which point we would not
need to discretise at all.

* Improve wording in panelling designer docstring

* Relax angle constraint in panelling length test

The minimum length was set quite high, especially for the cases using
just the in/outboard boundaries. This was making the tests very
sensitive. The idea of the test is just to check the lengths are
constrained, so relax the constraint on the angle, so the optimiser has
more valid space to work within.