EUDEMO blanket builder with panels (fusion-power-plant-framework/bluemira#2159)
* Rename misnamed test class

* Add panelling to eudemo blanket designer

* Use start point, not bbox, in ivc designer

Whilst trying to add panelling to the EUDEMO blanket segments, I ran
into a nasty-looking bug. Essentially, the 'cut_wall_boundary' wire
created within the 'design_ivc' function, is not self-consistent.
Dropping into a debugger on line 76 of eudemo/eudemo/ivc/__init__.py, I
can run the following:

  >>> cut_wall_boundary.bounding_box.z_min
  -5.701225441085458
  >>> cut_wall_boundary.start_point().z
  -5.52585334

This is nonsense, as the start of the wire is the lowest point of the
wire in z. Note that it is the bounding box value that is wrong here,
not the starting point value. The start point value is as we expect (the
same z as in the x-point we use to dictate the cut) and what is shown
when we plot the wire.

We then go on to use this incorrect bounding box value when we generate
the blanket face in `PlasmaFaceDesigner`. So the face we were generating
was actually 20cm too low!

There's some weird behaviour going on here. At the same point in the
debugger, I can now run this:

  >>> cut_wall_boundary.bounding_box.z_min
  -5.701225441085458
  >>> from bluemira.display import show_cad
  >>> show_cad([plasma_face, cut_wall_boundary], backend="POLYSCOPE")
  >>> cut_wall_boundary.bounding_box.z_min
  -5.525853343944378

Plotting the wire is somehow correcting the bounding box value. I
haven't been able to narrow down exactly what it is that causes the
'correction', as if I plot the wire immediately after creating it, the
correction does not occur.

In this commit, we switch to using the `start_point` method to get the z
coordinate with which to build the blanket face. This is a workaround,
rather than a solution.

* Fix some typos and typing in PlasmaFaceDesigner

* Add panelled blanket builder and add to eudemo

* Add comment on wire cutting to design_ivc

* Remove unused function from eudemo/reactor

* Add panelling params to eudemo params json

* Move eudemo blanket builder module

* Add tests for BlanketDesigner.segment_blanket

* Fix typo in eudemo.reactor

* Add angle size check to BlanketDesigner

* Add docstrings to BlanketDesigner

* Move panelling from ivc into blanket module

* Remove unused & incomplete PanellingBuilder class

This is no longer needed.

* Refactor BlanketDesigner.segment_blanket

* Add BlanketSegment class to blanket.designer

* Add eudemo blanket designer cutting test

* Remove TODO

This is to be resolved once the PR is merged.

* Move test_blanket into blanket test module

* Update eudemo blanket builder test

* Remove TODO

The function is actually used in one of the port builders, so it
probably doesn't belong in the blanket module.

* Fix incorrect unit in eudemo params JSON

* Improve 'fw_a_max' parameter description

* Refactor blanket testing utility into new file

We're going to need to use the function in another test file.

Also, move the blanket created by the utility down so that it crosses
the x-axis (i.e., it has area either side of z=0). This means we don't
get NoneType issues when calling `slice_shape` in
BlanketBuilder.build_xy using the xy-plane. Basically, the blanket must
cross z=0 for the xy part of the build to work correctly.

* Add eudemo_tests to known third-party in isort

* Add silhouette getters to Blanket comp. manager

* Improve robustness of eudemo clearance face func

* Reduce duplication in BlanketDesigner

* Hit blanket/test_designer.py with isort

* Fix incorrect BlanketDesigner method docstring